+++ tests/test_database.py
@@ -1,4 +1,3 @@
-import sqlite3
 import os
 import time
 import unittest
@@ -16,7 +15,7 @@ class DatabaseCreationTest(unittest.TestCase):
         os.remove("./database.db")
         if os.path.exists("./database.old.db"):
             os.rename("./database.old.db", "./database.db")
-            
+
     def test_database_creation(self):
         self.assertTrue(os.path.exists("./database.db"))
 
@@ -49,7 +48,7 @@ class DatabaseAccountTest(unittest.TestCase):
     def tearDown(self):
         if os.path.exists("./database.old.db"):
             os.rename("./database.old.db", "./database.db")
-        
+
     def test_create_user(self):
         self.database.create_user("jakob", b"memes")
         self.database.cursor.execute("SELECT * FROM users WHERE "