+++ tests/test_session.py
@@ -1,23 +1,29 @@
-# import unittest
+import unittest
 
-# from tests.dummy_objects import (DummyClient, DummyConfig, DummyDatabase)
-
-# from openbbs.session import (curry_io, handle)
-
-
-# class IOTest(unittest.TestCase):
-#     def test_curry_io(self):
-#         client = DummyClient(b"a")
-#         database = DummyDatabase()
-#         send, receive = curry_io(client, database, "")
-#         self.assertTrue(receive())
+from openbbs.config import load_config
+from openbbs.session import UserSession, handle
+from tests.dummy_objects import (DummyClient, DummyDatabase)
 
 
-# class ClientHandler(unittest.TestCase):
-#     def setUp(self):
-#         fakeconfig = DummyConfig()
-#         self.config_get = fakeconfig.fakeget
-#         self.client = DummyClient(b"quit")
+class UserEncapsulationTest(unittest.TestCase):
+    def test_basic_user(self):
+        client = DummyClient(b"a")
+        database = DummyDatabase()
+        user = UserSession(client, database, "")
+        self.assertTrue(user.receive())
+        with self.assertRaises(SystemExit):
+            user.close()
 
-#     def test_handler(self):
-#         self.assertFalse(handle(self.client, "0.0.0.0", self.config_get))
+
+class ClientHandler(unittest.TestCase):
+    def test_handler(self):
+        config = load_config("./inexistent.ini")
+        client = DummyClient(b"quit")
+        with self.assertRaises(SystemExit):
+            handle(client, "", config)
+
+    def test_banned_user(self):
+        config = load_config("./inexistent.ini")
+        client = DummyClient(b"login", b"a", b"a", b"quit")
+        with self.assertRaises(SystemExit):
+            handle(client, "", config)